requirements.txt
# Requirements para produção - Google Cloud Run
# Gerado automaticamente - não editar manualmente

# Django e dependências principais
Django>=4.2.7,<5.0
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
python-decouple>=3.8
whitenoise>=6.6.0

# Excel
openpyxl>=3.1.5

# PDF
reportlab>=4.0.0

# Outras dependências
Pillow>=10.0.0
django-extensions>=3.2.0







